move plot tab to leaderboard
